backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
shapely==2.0.5
networkx==3.3
numpy==1.26.4
python-multipart==0.0.12
orjson==3.10.7
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
